requirements.txt
google-generativeai==0.8.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
